requirements.txt
flask==2.3.3
werkzeug==2.3.7
twilio==8.10.0
groq==0.3.0
python-dotenv==1.0.0 